use gradle for building, reorg src/res dirs, add Sidekick lib source
just some "standardization" (debatable i guess) to the source directory
layout and resources directory.

removal of the dependancy on Sidekick-1.2 jar is mainly because Xunkar
does not have the source for this particular version of the library and
the current version on Github does not have some of the stuff that we
need for for this version (not even in an updated form, looks like he
did a fair bit of cleaning out of stuff he no longer needed). i would
rather _not_ depend on jars where no source / future support is possible
so i included decompiled sources for this library so future bugfixing
and modification is a bit easier moving forward. can decide if we want
to move to the latest version of Sidekick at a later time.