disabling language selection for now. can re-enable at a later date
quite simply, i don't speak anything other then english. the source
code i originally got from Xunkar also had some english strings that
had not been translated either, so everything other then english was
somewhat out of date to begin with.

i very strongly feel that it's bad form to provide an option for
choosing an alternate language within the app when the alternate
language translations are out of date and are just plain missing some
strings in places.

if at some point in the future someone wants to finish one or more of
these translations so they are once again in sync, we can then
re-enable the language selection.

it's unfortunate, but i feel this is the best decision for now.