remove resize options from the menu for now
i think these are probably somewhat confusing as-is, and i'd probably
need to change how they work somewhat with the new fixed-width window
option. easiest for now i think to just remove these menu options
as the window seems to be resizing automatically in a pretty good way
when the window autosize option is set. the only reason i can think of
that someone might want these options is to undo any manual window
resizings (done using the mouse, since the window is marked as user
resizable right now).