added separate flag for hotkey locking
already had an "ignore native inputs" flag which served a dual-purpose:
toggle flag for the menu's lock/unlock setting and also as a temporary
state that was set/unset when opening/closing the edit run and settings
dialogs. problem was that if you locked hotkeys, opened say the settings
dialog and then closed it, the hotkeys would be unlocked during the
dialog close event. this change fixes that behaviour.